Disabled binding to real IP, added javaopts flags
* Previous perfect scripts bound to an IP rather than localhost,
however neither 127.0.0.1 or an actual IP will work (only the localhost
alias, bizarre, i’d think at least 127.0.0.1 would work).  This may be
because the the “localhost” string is hard coded throughout the code
and may inadvertently have a comparison; which would fail if you use
127.0.0.1

* Added new JAVA_OPTS to perfect_install, the new java opts came from
rajesh and adds reunjdwp=transport, devlocal and debug flags (otherwise
it does not work on my machine).